diplom/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.1.0
drf-spectacular==0.26.5
django-rest-swagger==2.2.0
django-filter==23.4
drf-url-filters==0.5.1
environ==1.0
environs==9.5.0
celery==5.3.6
redis==5.0.1
